fix: frontend の Kotlin コンパイル警告を解消
CI に残っていた frontend の deprecation / 未使用 / opt-in 警告を排除:

- currentWindowAdaptiveInfo → currentWindowAdaptiveInfoV2(L/XL 幅対応版)
- Icons.Outlined.TrendingUp → Icons.AutoMirrored.Outlined.TrendingUp
- LocalDateTime.monthNumber/dayOfMonth → month.number/day
- 未使用の `Unit` 式を空ブロック `{}` に、冗長な null チェック / cast / toByte() を除去
- expect/actual class(Beta)を -Xexpect-actual-classes で抑制
- coroutines-test の ExperimentalCoroutinesApi を optIn
- wasmJs 相互運用(@JsFun 等)に @file:OptIn(ExperimentalWasmJsInterop)
- InviteSheet の LocalClipboardManager は suspend API 移行を伴うため
  当面 @Suppress("DEPRECATION") で抑制(挙動不変)

検証: :frontend の js/wasmJs main+test を --rerun-tasks で再コンパイルし
w:/e: 行ゼロ・BUILD SUCCESSFUL、spotlessCheck も green。

+++ frontend/src/jsMain/kotlin/net/brightroom/mindstock/frontend/auth/PkceJs.kt
@@ -8,7 +8,7 @@ import org.khronos.webgl.get
 
 internal actual fun secureRandomBytes(n: Int): ByteArray {
     val arr = createRandomBytesJs(n)
-    return ByteArray(n) { arr[it].toByte() }
+    return ByteArray(n) { arr[it] }
 }
 
 private fun createRandomBytesJs(n: Int): Uint8Array = js("globalThis.crypto.getRandomValues(new Uint8Array(n))")
@@ -17,7 +17,7 @@ internal actual suspend fun sha256(bytes: ByteArray): ByteArray {
     val input = Int8Array(bytes.toTypedArray())
     val digest = subtleDigestJs(input.buffer).await()
     val view = Uint8Array(digest)
-    return ByteArray(view.length) { view[it].toByte() }
+    return ByteArray(view.length) { view[it] }
 }
 
 internal actual fun base64UrlNoPad(bytes: ByteArray): String {
